requirements.txt
paho-mqtt==1.6.1
aiocoap==0.4.7
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
typer==0.12.5
uvloop==0.19.0
orjson==3.10.7
